Move to turns, rename to user_turn_completion_mixin.py; remove TYPE_CHECKING
