+++ src/pipecat/services/llm_service.py
@@ -56,7 +56,7 @@ from pipecat.processors.aggregators.llm_response import (
 from pipecat.processors.aggregators.openai_llm_context import OpenAILLMContext
 from pipecat.processors.frame_processor import FrameDirection
 from pipecat.services.ai_service import AIService
-from pipecat.services.mixins.turn_completion import TurnCompletionMixin
+from pipecat.turns.user_turn_completion_mixin import TurnCompletionMixin
 
 # Type alias for a callable that handles LLM function calls.
 FunctionCallHandler = Callable[["FunctionCallParams"], Awaitable[None]]
